requirements.txt
gradio==5.16.0
langchain==0.3.18
langchain-openai==0.3.5
langchain-chroma==0.2.1
matplotlib==3.10.0
scikit-learn==1.6.1
numpy==1.26.4
plotly==5.22.0
python-dotenv==1.0.1
